middleware/sub

+++ controllers/clothingitems.js
@@ -8,7 +8,9 @@ const getItems = async (req, res) => {
     return res.status(200).json(items);
   } catch (err) {
     console.error(err);
-    return res.status(SERVER_ERROR).json({ message: "Failed to retrieve items", error: err.message });
+    return res
+      .status(SERVER_ERROR)
+      .json({ message: "Failed to retrieve items", error: err.message });
   }
 };
 
@@ -17,7 +19,9 @@ const createItem = async (req, res) => {
   const owner = req.user._id;
 
   if (!name || !weather || !imageUrl) {
-    return res.status(BAD_REQUEST).json({ message: "All fields are required." });
+    return res
+      .status(BAD_REQUEST)
+      .json({ message: "All fields are required." });
   }
 
   try {
@@ -26,13 +30,12 @@ const createItem = async (req, res) => {
     return res.status(201).json(newItem);
   } catch (err) {
     console.error(err);
-    if (err.name === 'ValidationError') {
-      return res.status(BAD_REQUEST).json({ message: 'Invalid item data' });
+    if (err.name === "ValidationError") {
+      return res.status(BAD_REQUEST).json({ message: "Invalid item data" });
     }
-    return res.status(SERVER_ERROR).json({ message: "Failed to create item"});
+    return res.status(SERVER_ERROR).json({ message: "Failed to create item" });
   }
 };
-
 
 const deleteItem = async (req, res) => {
   const { itemId } = req.params;
@@ -42,14 +45,25 @@ const deleteItem = async (req, res) => {
   }
 
   try {
-    const deletedItem = await ClothingItem.findByIdAndDelete(itemId);
-    if (!deletedItem) {
+    const item = await ClothingItem.findById(itemId);
+    if (!item) {
       return res.status(NOT_FOUND).json({ message: "Item not found" });
     }
-    return res.status(200).json({ message: "Item successfully deleted", item: deletedItem });
+
+    const userId = req.user._id;
+    if (!item.owner.equals(userId)) {
+      return res
+        .status(403)
+        .json({ message: "You are not authorized to delete this item." });
+    }
+
+    const deletedItem = await ClothingItem.findByIdAndDelete(itemId);
+    return res
+      .status(200)
+      .json({ message: "Item successfully deleted", item: deletedItem });
   } catch (err) {
     console.error("Error occurred during deletion:", err);
-    return res.status(SERVER_ERROR).json({ message: "Failed to delete item"});
+    return res.status(SERVER_ERROR).json({ message: "Failed to delete item" });
   }
 };
